backend/requirements.txt
Django==4.2.7
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.0.0
elevenlabs>=1.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
websockets==12.0
djangorestframework==3.14.0
numpy==1.26.4
